Address review comment: rename Cloudflare cookie module
