+++ codex-rs/codex-client/src/lib.rs
@@ -1,4 +1,4 @@
-mod chatgpt_cookies;
+mod chatgpt_cloudflare_cookies;
 mod chatgpt_hosts;
 mod custom_ca;
 mod default_client;
@@ -9,7 +9,7 @@ mod sse;
 mod telemetry;
 mod transport;
 
-pub use crate::chatgpt_cookies::with_chatgpt_cloudflare_cookie_store;
+pub use crate::chatgpt_cloudflare_cookies::with_chatgpt_cloudflare_cookie_store;
 pub use crate::chatgpt_hosts::is_allowed_chatgpt_host;
 pub use crate::custom_ca::BuildCustomCaTransportError;
 /// Test-only subprocess hook for custom CA coverage.
